cat_db/requirements.txt
django == 2.2.7
djangorestframework == 3.10.3
factory_boy == 2.12.0
django-debug-toolbar == 2.1
